roll engine #cd34cdc6

+++ android/src/main/java/me/yohom/amap_map_fluttify/sub_handler/SubHandler16.java
@@ -108,9 +108,9 @@ public class SubHandler16 {
             
                     // args
                     // jsonable arg
-                    int var0 = (int) ((Map<String, Object>) args).get("var0");
+                    long var0 = (long) ((Map<String, Object>) args).get("var0");
                     // jsonable arg
-                    int var2 = (int) ((Map<String, Object>) args).get("var2");
+                    long var2 = (long) ((Map<String, Object>) args).get("var2");
             
                     // ref
             
@@ -144,7 +144,7 @@ public class SubHandler16 {
             
                     // args
                     // jsonable arg
-                    int var0 = (int) ((Map<String, Object>) args).get("var0");
+                    long var0 = (long) ((Map<String, Object>) args).get("var0");
                     // jsonable arg
                     float[] var2 = (float[]) ((Map<String, Object>) args).get("var2");
                     // jsonable arg
@@ -208,7 +208,7 @@ public class SubHandler16 {
             
                     // args
                     // jsonable arg
-                    int var0 = (int) ((Map<String, Object>) args).get("var0");
+                    long var0 = (long) ((Map<String, Object>) args).get("var0");
             
                     // ref
             
@@ -567,7 +567,7 @@ public class SubHandler16 {
             
                     // args
                     // jsonable arg
-                    int var0 = (int) ((Map<String, Object>) args).get("var0");
+                    long var0 = (long) ((Map<String, Object>) args).get("var0");
             
                     // ref
             
@@ -2136,7 +2136,7 @@ public class SubHandler16 {
             
                     // args
                     // jsonable arg
-                    int var1 = (int) ((Map<String, Object>) args).get("var1");
+                    long var1 = (long) ((Map<String, Object>) args).get("var1");
             
                     // ref
                     int refId = (int) ((Map<String, Object>) args).get("refId");
@@ -2171,7 +2171,7 @@ public class SubHandler16 {
             
                     // args
                     // jsonable arg
-                    int var1 = (int) ((Map<String, Object>) args).get("var1");
+                    long var1 = (long) ((Map<String, Object>) args).get("var1");
             
                     // ref
                     int refId = (int) ((Map<String, Object>) args).get("refId");
@@ -2206,7 +2206,7 @@ public class SubHandler16 {
             
                     // args
                     // jsonable arg
-                    int var1 = (int) ((Map<String, Object>) args).get("var1");
+                    long var1 = (long) ((Map<String, Object>) args).get("var1");
             
                     // ref
                     int refId = (int) ((Map<String, Object>) args).get("refId");
@@ -2276,7 +2276,7 @@ public class SubHandler16 {
             
                     // args
                     // jsonable arg
-                    int var1 = (int) ((Map<String, Object>) args).get("var1");
+                    long var1 = (long) ((Map<String, Object>) args).get("var1");
             
                     // ref
                     int refId = (int) ((Map<String, Object>) args).get("refId");
@@ -3149,7 +3149,7 @@ public class SubHandler16 {
             
                     // args
                     // jsonable arg
-                    int var1 = (int) ((Map<String, Object>) args).get("var1");
+                    long var1 = (long) ((Map<String, Object>) args).get("var1");
                     // ref arg
                     com.autonavi.amap.mapcore.animation.GLTransformation var3 = (com.autonavi.amap.mapcore.animation.GLTransformation) getHEAP().get((int) ((Map<String, Object>) args).get("var3"));
             
@@ -3187,7 +3187,7 @@ public class SubHandler16 {
             
                     // args
                     // jsonable arg
-                    int var1 = (int) ((Map<String, Object>) args).get("var1");
+                    long var1 = (long) ((Map<String, Object>) args).get("var1");
                     // ref arg
                     com.autonavi.amap.mapcore.animation.GLTransformation var3 = (com.autonavi.amap.mapcore.animation.GLTransformation) getHEAP().get((int) ((Map<String, Object>) args).get("var3"));
                     // jsonable arg
@@ -7116,9 +7116,9 @@ public class SubHandler16 {
             
                     // args
                     // jsonable arg
-                    int var0 = (int) ((Map<String, Object>) args).get("var0");
+                    long var0 = (long) ((Map<String, Object>) args).get("var0");
                     // jsonable arg
-                    int var2 = (int) ((Map<String, Object>) args).get("var2");
+                    long var2 = (long) ((Map<String, Object>) args).get("var2");
                     // jsonable arg
                     int var4 = (int) ((Map<String, Object>) args).get("var4");
