+++ android/src/main/java/me/yohom/amap_map_fluttify/sub_handler/SubHandler22.java
@@ -1704,7 +1704,7 @@ public class SubHandler22 {
             
                     // args
                     // jsonable arg
-                    int var1 = (int) ((Map<String, Object>) args).get("var1");
+                    long var1 = (long) ((Map<String, Object>) args).get("var1");
             
                     // ref
                     int refId = (int) ((Map<String, Object>) args).get("refId");
