+++ android/src/main/java/me/yohom/amap_map_fluttify/sub_handler/SubHandler9.java
@@ -6090,7 +6090,7 @@ public class SubHandler9 {
             put("com.amap.api.maps.offlinemap.OfflineMapProvince::setSize", (args, methodResult) -> {
                 // args
                 // jsonable arg
-                int var1 = (int) ((Map<String, Object>) args).get("var1");
+                long var1 = (long) ((Map<String, Object>) args).get("var1");
             
                 // ref
                 int refId = (int) ((Map<String, Object>) args).get("refId");
